+++ app/src/main/kotlin/com/aisleron/domain/product/Product.kt
@@ -25,7 +25,7 @@ data class Product(
     override val id: Int,
     override val name: String,
     val inStock: Boolean,
-    val qtyNeeded: Int,
+    val qtyNeeded: Double,
     override val noteId: Int? = null,
     override val note: Note? = null
 ) : AisleronItem, NoteParent
